Commit of my (possibly deprecated) ec2-new-instance scripts. Decided to go with chef/knife after all.

+++ bash/ec2/ec2-setup-eu-ireland.sh
@@ -4,10 +4,13 @@
 . ~/Repos/Scripts/bash/ec2/common/login.sh
 
 scp -i ec2-admin-eu-ireland.pem ~/Repos/Scripts/bash/ec2/common/first-run.sh ${ssh_target}:/tmp/first-run.sh
+scp -i ec2-admin-eu-ireland.pem ~/Repos/Scripts/bash/ec2/common/.gitconfig ${ssh_target}:~/.gitconfig
 
 tmpFile="/tmp/ec2-bash.sh"
 
-rm ${tmpFile}
+if [ -f ${tmpFile} ]; then
+    rm ${tmpFile}
+fi
 
 touch ${tmpFile}
 echo "cd /tmp" >> ${tmpFile}
